os_remote/test/Makefile has bugs
