+++ os_remote/test/remote_tool.cc
@@ -1,8 +1,10 @@
 #include "remote_tool.h"
 
+
 using std::string;
 
-void ConvertStringToBytes(const std::string & mess,char * bytes,,u32 bytes_len)
+
+void ConvertStringToBytes(const std::string & mess,char * bytes,u32 bytes_len)
 {
     mess.copy(bytes,bytes_len);
     return;
@@ -14,3 +16,4 @@ void ConvertBytesToString(const char * bytes,std::string & mess,u32 bytes_len)
     mess.swap(temp);
     return;
 }
+
