+++ os_remote/test/server_wrap.h
@@ -1,0 +1,6 @@
+#ifndef SERVER_WRAP
+
+int wrapOpen(char * vfs,const char * path,char * file,int in,int * out);
+
+#endif // SERVER_WRAP
+
